fix navbar
J'ai reparer la navbar en remplacant la page index par adoption et il n'y a plus de page index mais la page accueil.
